# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trading Buddy API",
    "description": "A股量化交易数据基础设施",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/factors/preview": {
      "get": {
        "tags": [
          "因子"
        ],
        "summary": "Factor Preview",
        "description": "对单列日 K 序列跑一项因子原语，不落库。",
        "operationId": "factor_preview_api_factors_preview_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "标的代码",
              "title": "Code"
            },
            "description": "标的代码"
          },
          {
            "name": "column",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "open",
                "high",
                "low",
                "close",
                "volume",
                "amount"
              ],
              "type": "string",
              "description": "K 线列；op=atr/kdj/cci/williams_r/mfi/obv/adx/aroon/donchian/vwap 时忽略（obv 用 close+volume；mfi/vwap 用 H/L/C/V；adx/aroon/donchian 用 H/L；可填 close）；roc/trix 与 rolling_* 等同按 column 取列",
              "title": "Column"
            },
            "description": "K 线列；op=atr/kdj/cci/williams_r/mfi/obv/adx/aroon/donchian/vwap 时忽略（obv 用 close+volume；mfi/vwap 用 H/L/C/V；adx/aroon/donchian 用 H/L；可填 close）；roc/trix 与 rolling_* 等同按 column 取列"
          },
          {
            "name": "op",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "rolling_sum",
                "rolling_mean",
                "rolling_std",
                "rolling_zscore",
                "rolling_max",
                "rolling_min",
                "ema",
                "pct_change_1",
                "pct_change_n",
                "diff_n",
                "rsi",
                "atr",
                "bollinger",
                "macd",
                "kdj",
                "cci",
                "obv",
                "williams_r",
                "mfi",
                "roc",
                "trix",
                "adx",
                "aroon",
                "donchian",
                "vwap"
              ],
              "type": "string",
              "description": "原语算子",
              "title": "Op"
            },
            "description": "原语算子"
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 500,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "rolling_*、ema、atr、bollinger、kdj、cci、williams_r、mfi、roc、trix、adx、aroon、donchian 等须传 window；vwap 滚动时须传 window（日级累计不传）；pct_change_n/roc/diff_n/rsi 为 n 或 period；trix 的 window 为 EMA span；adx/aroon 的 window 为 period；donchian 的 window 为通道宽度；pct_change_1、macd、obv 忽略",
              "title": "Window"
            },
            "description": "rolling_*、ema、atr、bollinger、kdj、cci、williams_r、mfi、roc、trix、adx、aroon、donchian 等须传 window；vwap 滚动时须传 window（日级累计不传）；pct_change_n/roc/diff_n/rsi 为 n 或 period；trix 的 window 为 EMA span；adx/aroon 的 window 为 period；donchian 的 window 为通道宽度；pct_change_1、macd、obv 忽略"
          },
          {
            "name": "bb_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 12.0,
              "minimum": 0.25,
              "description": "仅 op=bollinger：上下轨 k 倍 rolling_std；写入 meta.bb_k",
              "default": 2.0,
              "title": "Bb K"
            },
            "description": "仅 op=bollinger：上下轨 k 倍 rolling_std；写入 meta.bb_k"
          },
          {
            "name": "macd_fast",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "op=macd：快线 EMA span（须 < macd_slow）",
              "default": 12,
              "title": "Macd Fast"
            },
            "description": "op=macd：快线 EMA span（须 < macd_slow）"
          },
          {
            "name": "macd_slow",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "op=macd：慢线 EMA span",
              "default": 26,
              "title": "Macd Slow"
            },
            "description": "op=macd：慢线 EMA span"
          },
          {
            "name": "macd_signal",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "op=macd：DEA 对 DIF 的 EMA span；写入 meta.signal",
              "default": 9,
              "title": "Macd Signal"
            },
            "description": "op=macd：DEA 对 DIF 的 EMA span；写入 meta.signal"
          },
          {
            "name": "kdj_m1",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "description": "op=kdj：K 平滑参数 m1（默认 3）；写入 meta.m1",
              "default": 3,
              "title": "Kdj M1"
            },
            "description": "op=kdj：K 平滑参数 m1（默认 3）；写入 meta.m1"
          },
          {
            "name": "kdj_m2",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "description": "op=kdj：D 平滑参数 m2（默认 3）；写入 meta.m2",
              "default": 3,
              "title": "Kdj M2"
            },
            "description": "op=kdj：D 平滑参数 m2（默认 3）；写入 meta.m2"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 30,
              "description": "拉取日 K 根数上限（从新到旧再正序）",
              "default": 500,
              "title": "Limit"
            },
            "description": "拉取日 K 根数上限（从新到旧再正序）"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "response_format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "json",
                "csv"
              ],
              "type": "string",
              "description": "json：series/meta；csv：value / bollinger / macd / kdj / adx / aroon / donchian / vwap",
              "default": "json",
              "title": "Response Format"
            },
            "description": "json：series/meta；csv：value / bollinger / macd / kdj / adx / aroon / donchian / vwap"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}